Fixing problem norm_utils.h truncates to float unnecessarily

+++ src/core/calc/norm_utils.h
@@ -10,7 +10,7 @@
 
 namespace sketch2 {
 
-inline float compute_cosine_inverse_norm(const uint8_t* data, DataType type, size_t dim) {
+inline double compute_cosine_inverse_norm(const uint8_t* data, DataType type, size_t dim) {
     double norm_sq = 0.0;
     switch (type) {
         case DataType::f32: {
@@ -42,12 +42,12 @@ inline float compute_cosine_inverse_norm(const uint8_t* data, DataType type, siz
     }
 
     if (norm_sq == 0.0) {
-        return 0.0f;
+        return 0.0;
     }
-    return static_cast<float>(1.0 / std::sqrt(norm_sq));
+    return 1.0 / std::sqrt(norm_sq);
 }
 
-inline float compute_squared_norm(const uint8_t* data, DataType type, size_t dim) {
+inline double compute_squared_norm(const uint8_t* data, DataType type, size_t dim) {
     double norm_sq = 0.0;
     switch (type) {
         case DataType::f32: {
@@ -77,7 +77,7 @@ inline float compute_squared_norm(const uint8_t* data, DataType type, size_t dim
         default:
             throw std::runtime_error("compute_squared_norm: unsupported data type");
     }
-    return static_cast<float>(norm_sq);
+    return norm_sq;
 }
 
 } // namespace sketch2
