+++ src/core/storage/data_merger.cpp
@@ -71,9 +71,9 @@ Ret flush_and_close_merge_file(FILE** f, const char* context);
 float compute_stored_norm_for_dist(const uint8_t* data, DataType type, size_t dim, DistFunc dist_func) {
     switch (dist_func) {
         case DistFunc::COS:
-            return compute_cosine_inverse_norm(data, type, dim);
+            return static_cast<float>(compute_cosine_inverse_norm(data, type, dim));
         case DistFunc::L2:
-            return compute_squared_norm(data, type, dim);
+            return static_cast<float>(compute_squared_norm(data, type, dim));
         case DistFunc::DOT:
             throw std::runtime_error("compute_stored_norm_for_dist: DOT does not use stored norms");
         default:
